feat(aether-cli)!: add session usage and cost tracking (#405)
## Summary

- add shared token, model identity, pricing, and session usage wire
types
- track per-call and cumulative usage/cost across root agents,
sub-agents, compaction, and resumed sessions
- expose typed usage through agent events, ACP/headless output, session
analytics, telemetry, evals, and generated schemas
- decode structured tool display and sub-agent progress updates into
typed events

## Breaking change

`AgentEvent` usage payloads now use the new context and session usage
wire types. Existing TypeScript SDK releases cannot deserialize these
payloads and must be updated.

## Testing

- `just test` — 2720 passed, 15 skipped
- `git diff --check 69631171..HEAD`

+++ crates/llm/src/testing/llm_response.rs
@@ -33,7 +33,7 @@ impl LlmResponseBuilder {
         self
     }
 
-    pub fn usage(mut self, input_tokens: u32, output_tokens: u32) -> Self {
+    pub fn usage(mut self, input_tokens: u64, output_tokens: u64) -> Self {
         self.chunks.push(LlmResponse::usage(input_tokens, output_tokens));
         self
     }
